[WIP] Brackets builtin extension using Cameo Extension
This is work in progress builtin extension to enable Brackets run on
top of Cameo. We aim to make an external extension for this, so this
code doesn't need to be part of Cameo.

To run Brackets, you need to download the brackets code (but not
brackets-shell!) and use the workaround of HTML menus mentioned in

https://github.com/adobe/brackets/wiki/Linux-Version#re-enable-html-menus

Then build Cameo as usual and use the src/brackets/run-brackets.sh
script in Cameo repository to run it with the appropriated parameters.

The current code is able to open and save files. While we care about
the actual implementations of native functionality to work, the goal
here is to show native functionality being used in a real setting.