Update cameo.menu extension to new CameoExtension API
Leandro's commit was imported from another tree. This patch makes it
work on top of new code.

Major changes:
- Moved code from src/extensions, where we store the core extension
  subsystem, to src/menu directory;
- Make use of our existing context concept in CameoExtension;
- Post the work to be done in the main thread.

The goal of cameo.menu is to demonstrate that we can make extensions
that deal with UI features. In the current form it is useful for the
Brackets demo, but we intend to give it a better API and propose as an
extension to be part of Cameo runtime.