[Tools/Packaging] Allow installation of different apps in a system
Windows will not allow installation of an app if its product id is
the same as that of an installed app. Assigning product id to "*"
can solve this problem as wix will generate it automatically.
But this introduces another problem - the same apps will be
installed multiple times. So we have to implement some basic
upgrading mechanism, which requires configuring MajorUpgrade tag
and generating different UpgradeCode for differnt apps.