Implement cameo.menu for Brackets first
This commit changes cameo.menu API to fit whatever Brackets use as a
starting point. The previous code was aiming at a better API but it
ended up we would have more work to get Brackets itself running, because
it didn't provide the features needed by it (for instance, changing
title after the menu was created).

Once we finish Brackets demo, the plan is to evolve cameo.menu to be
more widely usable and ultimately be shipped with Cameo runtime. In
particular we would like to be able to implement the features exchanging
the least amount of messages as possible.